backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
google-genai>=1.5.0
python-dotenv>=1.0.1
pydantic>=2.9.0
redis>=5.0.0
google-cloud-firestore>=2.19.0
